Use miniforge to get python
We were using apt to get the system python earlier.
However, that is pinned to the distro, and can't be upgraded
separately. This is problematic for cases with reticulate
that want to use a newer python version.

Instead, we can get python from
[miniforge](github.com/conda-forge/miniforge/).
Miniforge is a community maintained conda installer that gives
us an independent, non-root installation of Python that is
separate from the system python. It also has a bunch of other
very useful characteristics that come from being a fully
separated and isolated environment than system python. I

This is what I now recommend people use over system python in
their own data science-y docker images. Felt like a good
time to put this upstream in rocker too!

- Remove pin of pip, since miniforge3 comes with a new enough pip
- Use CONDA_DIR to point to python base, not VENV_DIR